fix: move project related function to own subdir
