+++ main.go
@@ -1,11 +1,1 @@
 package tmuxproject
-
-import (
-	"github.com/jvzantvoort/tmux-project/config"
-)
-
-var (
-	mainconfig = config.NewMainConfig()
-)
-
-// vim: noexpandtab filetype=go
