+++ project/main.go
@@ -1,0 +1,11 @@
+package project
+
+import (
+	"github.com/jvzantvoort/tmux-project/config"
+)
+
+var (
+	mainconfig = config.NewMainConfig()
+)
+
+// vim: noexpandtab filetype=go
